Runner Performance Improvements: Non-Blocking Docker In Docker and Cached Base Image (#59)
Performance Improvements:
- We don't block on Docker daemon startup (see README in runner
directory) unless Docker is actually used.
- The base image is changed to one cached by Google Cloud Build. This
improves pull time by reducing the number of layers that need to be
pulled from the artifact registry before startup. The cached base images
do change as new versions come out, so we will need to automate updating
the base sha to ensure we don't regress.

Reliability Improvement:
- JITConfig is now gzipped before adding as a substitution. In testing I
had instances of it exceeding the 4,000 character limit for cloud build
substitutions. This was able to reduce the character count in my
testing.