Compatibility fixes (at the cost of performance) (#112)
Switch to a newer Ubuntu base image to get a newer version of libc. This
resolves linking issues with newer software, at the expense of no longer
caching the docker layer, and not using a google-built ubuntu image.

In the future we may remediate these performance losses by:
- Introducing image streaming
- Getting an updated base image into cloud build cache

Addresses but does not completely resolve #71 
Fixes #72 -- though I still want to find a better soultion