Implement Timeout For Runners That Don't Get Jobs (#110)
Currently if we schedule a runner but the job gets canceled before the
runner can pick it up, it will keep running until the cloud build
context ends. This is problematic for multiple reasons:

- When a new runner image is pushed, we don't know if new jobs being
scheduled are running on the new or old runner image, making validation
difficult.
- If a job gets scheduled and is picked up by a runner that is about to
exceed its cloud build context, the job could crash halfway through
processing.
- We pay the compute price for keeping the extra runner open. (Normally
not a huge issue, but problematic if a bug is causing lots of runners to
be started without jobs)

Additionally, this change enables two new possible features:
1. Spawning extra runners to clear out job queues in the event jobs are
scheduled but runner web hook fails for some reason. We can spawn extra
runners knowing if they don't pick up jobs they will not be leaked.
2. Pooling extra runners, to increase startup latency at the cost of
more compute. This requires expiration for the same reason any runner
that doesn't have a job does.

There is a race condition that is introduced:
There is some amount of time between a job being picked up by a runner
and the `pre-run.sh` script creating the lockfile. If the timeout
happens between these two times, then the job will be canceled. This is
inherent to the design of the actions runner, there is no completely
safe way to terminate a waiting runner.