Optional config to spawn multiple runners per webhook (#132)
In order to ensure that we don't get into the case that we have more
jobs than runners (which could happen if there is a temporary outage in
cloud build, cloud run, or github that causes webhook processing to
fail), we want to spawn multiple runners to clear out any backlog.

This is probably not a permanent solution. In the future I would rather
implement some sort of pooling mechanism or stuck job detection, but
this is a very simple implementation.

Excess jobs will be killed quickly due to #110.

Fixes #51 


Additional Fixes:
- Use logger.With() rather than appending a set of common log lines to
every context
- Remove a switch with a single case + default to an if statement to
reduce cyclomatic complexity