Ensure externals folder is copied over in the custom runner image (#158)
In an attempt to replace the github runner image with the custom one in
the ARC infrastructure, it was discovered that we must have the
externals folder available in the custom runner image.

There is an init container when running in `dind` mode that expects the
contents of this folder to be available.