New Runner Image Tagging Scheme (#157)
We are working on a new paradigm for managing runner tags.

Currently `latest` is always used in prodcution.

We will change this. Now we will have a more nuanced release cycle.

- In additional to tagging with :latest, we will tag runner with the git
sha
- `next` will be used as a staging point, promote-to-staging will allow
taking an existing image and tag it with `next`
- `cur` will be treated as our production version.
`promote-to-production` will take whatever image is in `next` and re-tag
it as `cur`
- `prev` will be used for rollbacks. when `cur` adopts `next`, `prev`
will adopt `cur` (if there are any changes)
-  This production will happen automatically every friday
- `cur` and `prev` can manually be set to arbitrary values if you use
the promote-to-production as a workflow dispatch. This helps in the case
where we need to have more control over deployed versions.