Allow additionally accepting self-hosted label along with configured label (#164)
This is a temporary feature to allow us to migrate away from using
self-hosted, as we need a state where both labels are accepted.

Once migration is complete, this should be deleted. This is why we
aren't putting a list of accepted labels, this boolean flag will be
easier to delete in the future.

In the future we will have a registration service that handles multiple
labels and target machine sizes.

Also:
- Remove dynamic `pr-` label/image tag as it was never used
- Enforce only single-tag jobs, as multi-tag jobs with a matching label
will never be scheduled on a runner (a runner must satisfy all tags in a
job, and we hard coded to only give a runner a single tag.